アクティビティ一覧部分のコンポーネント化
